Refactor file generation and diff extraction to include first commit
Changes version numbering so that the "ver0" file is a dummy blank file that will always get generated alongside the other files that correspond to versions of a file. This is so that during diff extraction and file comparison, the first commit will be included because it essentially adds to a blank file.